requirements.txt
streamlit
psycopg
sqlalchemy
psycopg-binary